requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
fake-useragent==0.1.11
frozenlist==1.3.1
idna==3.4
loguru==0.6.0
lxml==4.9.1
multidict==6.0.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
win32-setctime==1.1.0
yarl==1.8.1
